started working on LinkUpdaterImpl

+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/client/StackOverflowClient.java
@@ -4,5 +4,5 @@ import reactor.core.publisher.Mono;
 import ru.tinkoff.edu.java.scrapper.dto.response.response.StackOverflowQuestionResponse;
 
 public interface StackOverflowClient {
-    Mono<StackOverflowQuestionResponse> fetchQuestion(int questionId);
+    Mono<StackOverflowQuestionResponse> fetchQuestion(long questionId);
 }
